update so anyone can do sarthak clear command

+++ app.js
@@ -65,12 +65,14 @@ client.on('message', async msg => {
 
 		if (amount > 100) return msg.reply('You can`t delete more than 100 messages at once!'); // Checks if the `amount` integer is bigger than 100
 		if (amount < 1) return msg.reply('You have to delete at least 1 message!'); // Checks if the `amount` integer is smaller than 1
+		
+		if (message.author.id === '220352311422091264') {
+			msg.delete(); // deletes command message
+		}
 
-		await msg.delete(); // deletes command message
-
-		await msg.channel.fetchMessages({
+		msg.channel.fetchMessages({
 			limit: amount
-		}).then(async function (messages) {
+		}).then(function (messages) {
 				messages.forEach(message => {
 					if (message.author.id === '220352311422091264') {
 						// if user said it, then delete.
